backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
pandas==2.1.4
aiofiles==23.2.1
crewai==0.28.0
langchain>=0.1.10,<0.2.0
langchain-google-genai==0.0.6
tavily-python>=0.3.0 